Create en.rst
Add a document file in rst that I can link in the global documentation of SEPAL (https://github.com/openforis/sepal-doc). 
I find a way to dynamically integrate the documentation files. The only constraints is to be 1 page of pure .rst. 

You're module is the only one available in the dashboard that provide a one pager. So I take the liberty of translating it to .rst

When you'll accept my PR I'll be able to wire it to the full doc.